Updates required for the API to work correctly

+++ src/main/java/br/com/zup/softSkill/services/QuestionsService.java
@@ -1,5 +1,9 @@
 package br.com.zup.softSkill.services;
 
+import br.com.zup.infras.ResourceNotFoundException;
+import br.com.zup.softSkill.dtos.QuestionsRequestDTO;
+import br.com.zup.softSkill.dtos.QuestionsResponseDTO;
+import br.com.zup.softSkill.services.mappers.QuestionsMapper;
 import br.com.zup.softSkill.models.Questions;
 import br.com.zup.softSkill.repositories.QuestionsRepository;
 import org.slf4j.Logger;
@@ -8,8 +12,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
-import java.util.UUID;
+import java.util.stream.Collectors;
 
 @Service
 public class QuestionsService {
@@ -19,34 +22,44 @@ public class QuestionsService {
     @Autowired
     private QuestionsRepository questionsRepository;
 
-    public Questions createQuestion(Questions question) {
+    public QuestionsResponseDTO createQuestion(QuestionsRequestDTO request) {
         logger.info("Creating a new question");
-        return questionsRepository.save(question);
+        Questions question = QuestionsMapper.INSTANCE.toModel(request);
+        question = questionsRepository.save(question);
+        return QuestionsMapper.INSTANCE.toDto(question);
     }
 
-    public Questions updateQuestion(UUID id, Questions updatedQuestion) {
+    public QuestionsResponseDTO updateQuestion(Long id, QuestionsRequestDTO updatedRequest) {
         logger.info("Updating question with id: {}", id);
         return questionsRepository.findById(id).map(existingQuestion -> {
+            Questions updatedQuestion = QuestionsMapper.INSTANCE.toModel(updatedRequest);
             existingQuestion.setTitle(updatedQuestion.getTitle());
             existingQuestion.setOptionA(updatedQuestion.getOptionA());
             existingQuestion.setOptionB(updatedQuestion.getOptionB());
             existingQuestion.setOptionC(updatedQuestion.getOptionC());
-            return questionsRepository.save(existingQuestion);
-        }).orElse(null);
+            return QuestionsMapper.INSTANCE.toDto(questionsRepository.save(existingQuestion));
+        }).orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + id));
     }
 
-    public void deleteQuestion(UUID id) {
+    public void deleteQuestion(Long id) {
         logger.info("Deleting question with id: {}", id);
+        if (!questionsRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Question not found with id " + id);
+        }
         questionsRepository.deleteById(id);
     }
 
-    public List<Questions> listQuestions() {
+    public List<QuestionsResponseDTO> listQuestions() {
         logger.info("Listing all questions");
-        return questionsRepository.findAll();
+        return questionsRepository.findAll().stream()
+                .map(QuestionsMapper.INSTANCE::toDto)
+                .collect(Collectors.toList());
     }
 
-    public Questions getQuestionById(UUID id) {
+    public QuestionsResponseDTO getQuestionById(Long id) {
         logger.info("Getting question with id: {}", id);
-        return questionsRepository.findById(id).orElse(null);
+        return questionsRepository.findById(id)
+                .map(QuestionsMapper.INSTANCE::toDto)
+                .orElseThrow(() -> new ResourceNotFoundException("Question not found with id " + id));
     }
 }
